requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiogram==3.2.0
motor==3.6.0
pydantic==2.5.2
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
httpx==0.25.2
